+++ internal/preset/scaffold_rename_linux.go
@@ -1,0 +1,9 @@
+//go:build linux
+
+package preset
+
+import "golang.org/x/sys/unix"
+
+func renameScaffoldBundle(from, to int, name string) error {
+	return unix.Renameat2(from, "bundle", to, name, unix.RENAME_NOREPLACE)
+}